Create Move puppy to left.png
Didn't like how the gradient text looked. I was going for the text as the bars being the last thing you noticed so i don't mind that it blends a bit/takes a bit more effort to read. But when I moved the puppy to the left it made it easier to see the text on the right ear